rmem: memory stats implemented
When CONFIG_RINA_MEMORY_STATS, the rmem module will
keep track of the number of allocated RINA objects. The objects
are classified using their size (approximated to a power of 2).
The current state is logged every 2 seconds.

The script rina-tools/mem-stats.py can be used to automatically
extract and show this information.

Signed-off-by: Vincenzo Maffione <v.maffione@nextworks.it>