Ignore SIGINT in IPC processes
When debugging the IPCM in gdb, if SIGINT is not ignored, a Ctrl+C in the
debugger signals the child process and, since it is not captured, it dies.

This patch sets SIGINT to be ignored by IPCPs.

This patch requires commit 698eb30198226ac4b3e3d264af1183a5206a7fd1 PR to be
integrated first.