Reverted API change on *::add_obj/addRIBObj
As discussed in bisdn/irati#29, the API to add objects to a RIB
(RIBDaemonProxy, RIBDaemon and RIB) were deliverately using a double pointer,
to set to NULL the variable holding the pointer and therefore make it clear
that the memory pointed is now managed by the library.

This commit reverts this change and restores the lost test assertions.