rinad & kernel: ipcp deamon calculates RED-ps LUT
rinad:ipcp
kernel:red-ps
The RED's ps lookuptable is calculated (floating point operations) in
IPCProcessImpl::processAssignToDIFRequestEvent. It is an adhoc code for the case
of the "red-ps". The address of the LUT is passesd as a string to the kernel
and down there the LUT is recovered by copy_from_user.
Still to decide the best place (within the ipcp) for this patch.

kernel:rmt
rmt_config needs to to be kept because it contains also the policy set configuration
which may be needed by a PS when being selected. For this reason too a rmt_config
getter is added.