kernel: cas-ps: clean up of #703 and fix LKM compilation
After the changes in #703 the w_dec_b_num and w_dec_b_den were not needed
since the parameter is now not configurable.

The division of the average length of the queue in the rmt-cas-ps.c avoided
the module to compile due to a division of (unsigned long) / (signed long long) that was not
supported.