kernel: cas-ps: rmt-ps: using absolute value of the time delta and casting it to ulong.
For some reason the / operator can not divide by a s64 variable in some machines. A
negatime time delta is not expected and an error. An error message is logged if this
happens and the s64 value of the time delta is always passed through abs64() and casted
to ulong.