kernel: dtcp: check ps->rtt_estimator is not NULL
Nothing prevents a PS to not define the rtt_estimator op and set it
to NULL. If this happens, the DTCP code executes the op without checking
if the pointer is not NULL and the system crashed.
The proper solution would be to check for mandatory ops in the PS, but in
the meanwhile this solves the problem.