kernel:red-ps: added RMT debug proc entry
Following the model usd in DTCP, the RMT component now logs to file
in procfs. It only supports 1 file per system so only 1 ipcp in the
system using red-ps.

Both RMT and DTCP debug code is not compiled if the flag is disabled
in the corresponding rmt-ps-debug.h and dtcp-ps-debug.h.