kernel: shim-eth-vlan: separated qdisc handling code from the shim-eth-vlan
In order to allow qdisc utils to be re-used by other shim-eths, they have been
moved out to a separate file, shim-eth-utils.c. A header file with declarations
of the required operations has also been added shim-eth.h, and shim-eth-vlan.c
had to be renamed to shim-eth-vlan-main.c. This is becasue the shim-eth-vlan
kernel module now has multiple compilation units (shim-eth-vlan.o and shim-eth-
utils.o), and the name of the kernel module cannot be the same as the name
of one of its compilation units, therefore now the modules is still called
shim-eth-vlan and has shim-eth-vlan-main.o and shim-eth-utils.o as compilation
units (thanks to Vincenzo for the explanation).

When we add another shim-eth we'll have to slightly modify shim-eth-utils to
accept a function pointer, since each shim-eth-x will have a different
implementation of the enable-write-all function.