kernel-rinad:red-ps:RED omits the use of idle time
In order to compute average queue usage, RED uses an approximation
when the idle period is over (queue was empty and a new PDU arrived).
But doing this the qavg calculation seems to be distorted. For this
PS we can afford to calculate the avg also when pdus are dequeued
using the formula qavg = qavg(1-W) + qW.

This change makes the calculation of the LUT introduced in addfb77
not needed anymore. For this reason the changes in rinad:ipcp are
reversed since it was a bit of a hack.