kernel:efcp: fix #778 waitqueue to destroy flow
Originally in the receive/send operations of the efcp_container if the
pending_ops counter arrived to 0 and the EFCP instance in action was in
EFCP_DEALLOCATED state, efcp_destroy would be called. This would call
dtp/dtcp_destroy which would do a rina_component_fini. This op takes a mutex.
The problem is that all this can happen in interrupt context.

The solution implemented here is to use a wait queue in the
efcp_connection_destroy operation (always called in process context) which sets
the state of efcp instance to EFCP_DEALLOCATED. After doing it, it sleeps till
the pending_ops counter is 0.  The receive/send operations now wake_up the
efcp_connection_destroy instead of calling efcp_destroy directly.