kernel: add default implementation of the SDU Protection policy set
This adds the default implementation of the SDU Protection policy set to the
default plugin module. It's based on how the default policy set is implemented
for the RMT. The HMAC and Compression hooks are just empty functions for now
since they require changes on the user space side. Other policies are extracted
from the PoC implementation.