kernel: normal-ipcp: add SDU Protection module
This patch adds the SDU Protection module to the normal ipcp process. We also
give it's reference to the RMT module so that it can call it directly. All the
SDU Protection related calls going through the normal ipcp API are now
redirected into the SDU Protection module instead of the RMT. Though since SDU
Protection is not connected to the RMT yet, the SDU Protection mechanism won't
be called.