MA-Manager included DIF configurations
The MA and the Manager have been modified to include DIF configurations
in the IPC process creation (assignment phase). Now, the MA does not use
a template with the dif configuration but receives all the configurations
from the manager including the DIF configurations but also the neighbors
to enroll.

To this intent, several files have been modified to achieve:
1. Cretation (or reuse) of the configuration structures
2. Modification of the assign to dif process to allow not only a template
but also a direct configuration call
3. Modification of the MA functionalities in the IPC process creation
4. Creation of the needed encoders.