kernel:robjects: sysfs-utils moved to robjects.h
The previous sysfs-util.h file was moved to rds/robjects.h. It contains the
sysfs macros to declare the required attributes for the kobjects and a
minimimalistic set of wrappers to the only kobject_* APIs used. The wrappers
implement some workarounds needed for sanity and ease some operation that would
have required more than 1 line. It still relies 100% on kobjects though.