Gracefully closing of the read mng thread
Fixes #825
Added error codes on the kernel side that are caught in the userspace.
When a critical error occurs on the read of the mng sdu in the kernel
it means that the kernel side of the normal IPCP is being closed.
In this case the userspace stops the thread that is polling
the kernel