kernel: Create ARP table upon adding entry
This will create a table for a specific device when arp826 is asked to
add a GPA entry, in case it was not created yet. Previously it would
bail out if the table did not exist yet. This will allow one to load
the shim-eth-vlan module even after adding a VLAN interface.

Fixes #242