librina: security-manager: add mac key generation
This commit adds key generation for the MAC policies. We generate two different
keys (server and client) that are then used as rx/tx (tx/rx) keys by the
enroller(enrollee). Keys are generated based on RFC4253.

This commit also adds the option to used the SSL_TXT_SHA256 MAC algorithm in
addition to the MD5 and SHA1 algorithms.