kernel: sdup-crypto-ps-default: add sequence numbers and replay detection
This adds sequence numbers to the beginning of the protected SDUs. While
receiving a protected SDU we also do some form of replay protection by using a
bitmap of previously received SDU sequence numbers. This is the same approach as
is used in IPSec. Users can configure the sequence number window size that
determines the oldest acceptable sequence number.