kernel: buffer: export buffer_assign function
This function will be used by the SDU Protection crypto policy set, in
particular the compression policy.
Even though the function is not used anywhere outside the buffer module and is
marked for removal/heavy rework, nfortunately compression isn't guaranteed to
happen in-place and can result in a bigger output size than the original it is
easier to replace the original buffer. This is where the buffer_assign function
comes in.