kernel: sdup-crypto-ps-default: add compression
This adds the finall SDU Protection crypto policy - compression. It can be
configured with the "compressionAlg" policy parameter, that currently accepts
the value "deflate". Compression is done out-of-place and uses a preallocated
scratch buffer that is created during the initialization of the sdu protection
private data structure. This approach is inspired by how IPSec does compression
within the kernel.