Delegation of objects various bugs fixed
Flags where not being encoded as expected. Delegation objects are now
being sent to the manager controling the invoke_id used as response and
also the port of the conexion (which was not being stored, and hence,
not allowing multiple managers). It also deals with the race condition
generated by the delegation on the last object sent to the manager
(with no flags) that finishes the flow of cdap objects. To control
this problem, the MA will not sent its last RIB object until all
the delegated objects have finished their work and hence, all the
delegated objects have been send by the MA to the manager. Then,
the MA will send the last response object of the cdap flow for a
given invoke_id
Fixes #932
Fixes #934