kernel:sk_buffer integration main commit
This commit implements the new PCI/SDU/PDU structures based on the addition of
an sk_buff. This implementation may not be the final one, huge effort was done
to conserve most of current APIs untouched (mostly pci.h).
A new struct du is added as unique data unit model. pdu and sdu are wrappers
over this struct in order to maintain existing apis.
struct pci modified contain a pointer to the position in the current du's buffer
where the pci is located.

Most of the kernel components have been refactored to the new model and APIs.
These are:

 - syscalls
 - kipcm
 - ipcp-instances
 - normal ipcp
 - efcp
 - dt-utils
 - dtp
 - dtcp
 - rmt
 - kfa

 - serdes removed