Added java catch for exceptions thrown in ipc-api
If exceptions are not explicetly declared in the .h function
declaration, SWIG does not wrap them and hence, they can not be
catched in java. To avoid the problem without changing the .h,
it is possible to define the catch in the .i.
This commit wraps all the catches thrown by functions in
librina/ipc-api.h
Fixes #947