mad: Produce IPCProcess ID
The id of the ipcprocess was being sent by the manager, which is a
wrong behaviour. This characteristic produced also some problems
when erasing the created IPCP etc... Now the original message is a
CREATE(".../IPCProcessID") instead of CREATE(".../IPCProcessID=x").
Now the MAD is able to remove the rib objects when IPCProces dies.

Fixes #952