pdu:added PDU_HEAD_GROW_WITH_PCI in grow/shrink
Added define to modify the behaviour in pdu_head_grow and pdu_head_shrink
functions. When defined PCI is not considered (as if PDU is and SDU). This is
the default behaviour since these functions are only used in SDUP when the PDU
is just to be passed down or received from IPCP below.