Alternate syntax for shim-tcp-udp dirEntry and expReg config entries
Add parsing of an additional syntax for config entries.  It is no
longer necessary to count characters in the fields.  To smooth the
transition to the new syntax, the old syntax is still accepted.

In the new syntax, every field begins with ':' including the initial
count of the number of entries.  This distinguishes new-syntax
config strings from old ones, as old configuration strings begin with
a number and new ones with a colon.  Old syntax:
	entries:count:fieldcount:fieldcount:fieldcount:field
  vs. new syntax:
	:entries:field:field:field:field

Example:
old:	"dirEntry" : "1:9:razzycam40:13:192.168.1.1154:2426",
	"expReg" : "1:9:razzycam20:4:2426"

new:	"dirEntry" : ":1:razzycam4::192.168.1.115:2426",
	"expReg" : ":1:razzycam2::2426"

The meaning of the "entries" field above, which is the number of
entire configuration entries in the field, is unchanged.  The remaining
fields are repeated with no intervening punctuation.  Example:
	"dirEntry" : ":2:razzycam4::192.168.1.115:2426:razzyr::192.168.1.149:2426",

The colon is now an illegal character in any field, but no other
differences in the resulting kernel config data result from this change.

Other changes:
Moved common code in parse_assign_conf() into a new common function
get_syntax_and_count() that also detects the new vs. old syntax
and takes care of freeing a kernel copy in event of an error.

Fixed latent bugs in detection of illegal entry count.

Added additional INFO logfile output to aid in debugging configurations.