Initialize the 'timed_out' value to something
I've been seeing the error message "ARP resolver called on timed out request"
without realizing that it meant that something was quite wrong in the code. It
turns out that when an ARP reply was received, the value of "timed_out" was not
initialized, which meant proper replies got rejected on a regular basis. Enough
requests still went through as normal that I didn't realize something was wrong
before now.